Reorganizacion de entitys
