+++ src/persistence/entities/account-type.entity.ts
@@ -1,6 +1,6 @@
 import { AccountTypeModel } from '../../models';
+import { GeneralCRUD } from '../repositories';
 import { v4 as uuid } from 'uuid';
-import { GeneralCRUD } from '../repositories/base/GeneralCRUD.base';
 
 export class AccountTypeEntity extends GeneralCRUD<AccountTypeModel> implements AccountTypeModel{
     id = uuid();
