+++ src/persistence/entities/document-type.entity.ts
@@ -1,6 +1,6 @@
 import { DocumentTypeModel } from '../../models';
-import { v4 as uuid } from 'uuid';
 import { GeneralCRUD } from '../repositories';
+import { v4 as uuid } from 'uuid';
 
 export class DocumentTypeEntity extends GeneralCRUD<DocumentTypeModel> implements DocumentTypeModel{
     id = uuid();
