+++ src/persistence/repositories/customer.repository.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
+import { GeneralCRUD } from './base';
 import { CustomerEntity } from '../entities';
-import { GeneralCRUD } from './base/GeneralCRUD.base';
 
 @Injectable()
 export class CustomerRepository extends GeneralCRUD<CustomerEntity> {
