+++ src/persistence/repositories/transfer.repository.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
+import { GeneralCRUD } from './base';
 import { TransferEntity } from '../entities';
-import { GeneralCRUD } from './base/GeneralCRUD.base';
 
 @Injectable()
 export class TransferRepository extends GeneralCRUD<TransferEntity> {
